+++ renpy/gl/glenviron_fixed.pyx
@@ -319,7 +319,7 @@ cdef class FixedFunctionEnviron(Environ):
         else:
             glDisableClientState(GL_TEXTURE_COORD_ARRAY)
 
-    cdef void set_color(self, double r, double g, double b, double a):
+    cdef void set_color(self, float r, float g, float b, float a):
         glColor4f(r, g, b, a)
 
     cdef void ortho(self, double left, double right, double bottom, double top, double near, double far):
